requirements.txt
openai
pandas
numpy
PyPDF2
streamlit
requests
json
python-dotenv